fatto cose

+++ proc_function.h
@@ -7,6 +7,6 @@ int take_time(char** buffer);
 char* take_res(char** buffer);
 char* take_share(char** buffer);
 char* take_cmdline(char* path, char* string);
-float take_cpu_usage_process(char** buffer_stat);
+double take_cpu_usage_process(char** buffer_stat);
 
 int search_group_id(char* path);
